market_data_service/requirements.txt
fastapi==0.68.0
uvicorn==0.15.0
sqlalchemy==1.4.22
aiocache==0.11.1
redis==4.0.2
yfinance==0.1.63
python-socketio==5.11.2